03-developer-journey/requirements-for-agentcore.txt
strands-agents[otel]>=1.21.0
strands-agents-tools>=0.2.19
boto3>=1.42.25
bedrock-agentcore-starter-toolkit>=0.2.5
langfuse>=3.11.2
python-dotenv>=1.2.1
duckduckgo-search>=8.1.1
pydantic>=2.0.0
